ToolsStation/backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
lxml
python-dotenv
loguru
httpx
wechatpy[cryptography]
apscheduler
sqlalchemy
openai
newspaper3k
lxml_html_clean
